refactor: unify logic for user hook
