--- src/ui/mobile/hooks/useUser.ts
@@ -1,11 +1,0 @@
-import { useCallback } from "react";
-
-import { updateAvatarUseCase } from "@/application/users/updateAvatarUseCase";
-import type { UserAvatar } from "@/domain/entities/Avatar";
-import { userRepository } from "@/infrastructure/convex/repository/userRepository";
-
-export function useUpdateAvatar() {
-  return useCallback(async ({ avatar }: { avatar: UserAvatar }) => {
-    return await updateAvatarUseCase(userRepository, avatar);
-  }, []);
-}
