--- src/ui/shared/login/useCheckEmailExists.ts
@@ -1,19 +1,0 @@
-import { useState } from "react";
-
-import { checkEmailExistsUseCase } from "@/application/users/checkEmailExistsUseCase";
-import { userRepository } from "@/infrastructure/convex/repository/userRepository";
-
-export const useCheckEmailExists = () => {
-  const [isChecking, setIsChecking] = useState(false);
-
-  const checkEmailExists = async (email: string) => {
-    setIsChecking(true);
-    try {
-      return await checkEmailExistsUseCase(userRepository, email);
-    } finally {
-      setIsChecking(false);
-    }
-  };
-
-  return { checkEmailExists, isChecking };
-};
