+++ src/ui/web/modules/settings/components/DisplayNameSection.tsx
@@ -3,6 +3,11 @@ import { type ChangeEvent, useEffect, useMemo, useState } from "react";
 import { CircleCheck, CircleX, Loader } from "lucide-react";
 import { useDebouncedCallback } from "use-debounce";
 
+import {
+  useCheckUsernameExists,
+  useCurrentUser,
+  useUpdateUsername,
+} from "@/ui/shared/user/hooks/useUser";
 import { Small } from "@/ui/web/components/Typography";
 import { Button } from "@/ui/web/components/ui/button";
 import {
@@ -10,11 +15,6 @@ import {
   InputGroupButton,
   InputGroupInput,
 } from "@/ui/web/components/ui/input-group";
-import {
-  useCheckUsernameExists,
-  useCurrentUser,
-  useUpdateUsername,
-} from "@/ui/web/hooks/useUser";
 
 export const DisplayNameSection = () => {
   const currentUser = useCurrentUser();
